add password reset timeout

+++ src/server/ts/domain/user/_service/user-password.ts
@@ -1,7 +1,7 @@
 import { NewPasswordForm } from '@common/type/password'
 import { userCheckPassword } from '@server/domain/user/_service/_user-check'
 import { Mailer } from '@server/mailer/mailer2'
-import { EMAIL_NOT_FOUND, EMAIL_PATTERN, INVALID_DATA } from '@common/type/error-const'
+import { EMAIL_NOT_FOUND, EMAIL_PATTERN, INVALID_DATA, PASSWORD_RESET_TIMEOUT } from '@common/type/error-const'
 import { emailPatternIsOk } from '@common/util/email'
 import { randomBytes } from 'crypto'
 import { ErrorConst } from '@common/type/error'
@@ -9,11 +9,11 @@ import { PwMailDB } from '@server/db/password/pwmail-db'
 import { makeHash } from '@common/util/hash'
 import { UserDB } from '@server/db/user/user-db'
 import { omanGetConfig } from '@server/oman/oman'
+import { dateDiffMins } from '@common/util/date2'
 
-export async function userSendPasswordMail(
+export async function userPasswordMail(
   mailer: Mailer, udb: UserDB, rdb: PwMailDB, email: string, err: ErrorConst[]
 ) {
-
   if (!emailPatternIsOk(email)) {
     err.push(EMAIL_PATTERN)
     return false
@@ -27,9 +27,10 @@ export async function userSendPasswordMail(
 
   const id = rdb.getNextId()
   const random = randomBytes(16).toString('hex')
+  const cdate = new Date()
 
   await rdb.deleteByEmail(email)
-  await rdb.insert({ id, email, random })
+  await rdb.insert({ id, email, random, cdate })
 
   const config = omanGetConfig()
   const mail = {
@@ -39,18 +40,16 @@ export async function userSendPasswordMail(
     text:
       '\n' +
       'Open the following URL to reset your password.\n\n' +
-      config.mainUrl + '/user-password-reset?id=' + id + '&r=' + random + '\n\n' +
+      config.mainUrl + '/user-password-reset/' + id + '/' + random + '\n\n' +
       config.appName
   }
 
   return mailer.sendMail(mail)
-
 }
 
-export async function userResetPassword(
+export async function userPasswordReset(
   udb: UserDB, rdb: PwMailDB, form: NewPasswordForm, err: ErrorConst[]
 ) {
-
   const id = form.id
   const random = form.random
   const password = form.password
@@ -69,11 +68,14 @@ export async function userResetPassword(
     err.push(INVALID_DATA)
     return
   }
+  if (dateDiffMins(new Date(), r.cdate) > 5) {
+    err.push(PASSWORD_RESET_TIMEOUT)
+    return
+  }
 
   const email = r.email
   const hash = await makeHash(password)
-
   await udb.updateHash(email, hash)
-  await rdb.deleteByEmail(email)
 
+  await rdb.deleteByEmail(email)
 }
